1、产线在制队列汇总查询修改 2、看板设备维护修改 3、产线资源组配置修改 2019.7.5 918100064

+++ core/src/main/java/yj/kanb/vblinegroupheader/service/impl/VblinegroupheaderServiceImpl.java
@@ -5,9 +5,12 @@ import com.hand.hap.system.service.impl.BaseServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import yj.kanb.vbgroupdtl.dto.Vbgroupdtl;
+import yj.kanb.vbgroupdtl.mapper.VbgroupdtlMapper;
 import yj.kanb.vblinegroupheader.dto.Vblinegroupheader;
 import yj.kanb.vblinegroupheader.mapper.VblinegroupheaderMapper;
 import yj.kanb.vblinegroupheader.service.IVblinegroupheaderService;
+import yj.kanb.viewdataschemaline.mapper.ViewdataschemalineMapper;
 
 import java.io.IOException;
 import java.net.URL;
@@ -19,6 +22,10 @@ import java.util.*;
 public class VblinegroupheaderServiceImpl extends BaseServiceImpl<Vblinegroupheader> implements IVblinegroupheaderService {
     @Autowired
     private VblinegroupheaderMapper vblinegroupheaderMapper;
+    @Autowired
+    private VbgroupdtlMapper vbgroupdtlMapper;
+    @Autowired
+    private ViewdataschemalineMapper viewdataschemalineMapper;
     @Override
     public List<Vblinegroupheader> selectAllGroup() {
         return vblinegroupheaderMapper.selectAllGroup();
@@ -30,12 +37,19 @@ public class VblinegroupheaderServiceImpl extends BaseServiceImpl<Vblinegrouphea
     }
 
     @Override
-    public void deleteLineGroupH(IRequest requestCtx, List<Vblinegroupheader> dto) {
+    public String deleteLineGroupH(IRequest requestCtx, List<Vblinegroupheader> dto) {
         if(dto.size() >0){
             for(int i=0;i<dto.size();i++){
-                vblinegroupheaderMapper.deleteLineGroupH(dto.get(i));
+                List<Vbgroupdtl> list = vbgroupdtlMapper.queryByGroupId(dto.get(i).getGroupId());
+                if (list.size() > 0){
+                    return "该产品已配置产线，不能删除!";
+                }else{
+                    viewdataschemalineMapper.deleteViewdataschemaline(dto.get(i).getGroupId());
+                    vblinegroupheaderMapper.deleteLineGroupH(dto.get(i));
+                }
             }
         }
+        return null;
     }
 
     @Override
